backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
Pillow==10.4.0
easyocr==1.7.1
